Add assets to transformation to make fixtures happy
