+++ src/elements/image/imageElementDataTransformer.ts
@@ -1,8 +1,9 @@
 import pickBy from "lodash/pickBy";
 import type { FieldNameToValueMap } from "../../plugin/helpers/fieldView";
+import type { Asset } from "../helpers/defaultTransform";
 import { undefinedDropdownValue } from "../helpers/transform";
 import type { TransformIn, TransformOut } from "../helpers/types/Transform";
-import type { Asset, createImageFields, MainImageData } from "./ImageElement";
+import type { createImageFields, MainImageData } from "./ImageElement";
 
 export type ImageFields = {
   alt?: string;
